requirements.txt
# Core Flask and Web Components
Flask
Flask-SQLAlchemy
Flask-Migrate
Flask-Login
Flask-WTF
python-dotenv
werkzeug
email-validator

# Background Task Processing
celery
redis
gevent

# Web Scraping
playwright

# Utilities
gunicorn
psycopg2-binary